+++ apps/admin/app/page.tsx
@@ -3,11 +3,7 @@ import { SessionProvider } from "next-auth/react";
 import ReactAdmin from "./ReactAdmin";
 
 const HomePage = () => {
-  return (
-    <SessionProvider>
-      <ReactAdmin />
-    </SessionProvider>
-  );
+  return null;
 };
 
 export default HomePage;
