+++ apps/admin/package.json
@@ -11,23 +11,23 @@
   },
   "dependencies": {
     "@babel/core": "7.23.7",
-    "@mui/icons-material": "^5.15.11",
-    "@mui/material": "^5.15.11",
-    "@mui/styles": "^5.15.11",
-    "axios": "^1.6.7",
+    "@mui/icons-material": "^5.15.14",
+    "@mui/material": "^5.15.14",
+    "@mui/styles": "^5.15.14",
+    "axios": "^1.6.8",
     "db": "workspace:*",
     "lodash": "^4.17.21",
     "next": "14.0.4",
-    "next-auth": "^4.24.6",
+    "next-auth": "^4.24.7",
     "next-auth-prisma-adapter": "workspace:*",
     "ra-data-simple-prisma": "workspace:*",
     "react": "18.2.0",
-    "react-admin": "^4.16.11",
+    "react-admin": "^4.16.13",
     "react-admin-json-view": "^2.0.0",
     "react-dom": "18.2.0"
   },
   "devDependencies": {
-    "@types/lodash": "^4.14.202",
+    "@types/lodash": "^4.17.0",
     "@types/node": "20.10.6",
     "@types/react": "18.2.46",
     "@types/react-dom": "18.2.18",
