+++ packages/ra-data-simple-prisma/package.json
@@ -38,17 +38,16 @@
     "@prisma/client": ">=3"
   },
   "dependencies": {
-    "axios": "^1.5.0",
+    "axios": "^1.6.8",
     "deepmerge": "^4.3.1",
-    "deverything": "^0.28.1",
-    "set-value": "^4.1.0"
+    "deverything": "^0.48.0"
   },
   "devDependencies": {
-    "@types/node": "^20.7.0",
+    "@types/node": "^20.11.30",
     "eslint-config-custom": "workspace:*",
     "jest": "^29.7.0",
-    "react-admin": "^4.14.3",
+    "react-admin": "^4.16.13",
     "tsup": "^7.2.0",
-    "typescript": "^5.2.2"
+    "typescript": "^5.4.3"
   }
 }
